Drivetrain, FlagWaver, OI/IO Changes Fixed the drivetrain code by adding Phoenix library.
Also improved FlagWaver commands by seperating the start and stop into
two seperate commands, which allows for it to be used in OI button
mapping. Previously, the command used conditional statements which was
bad for coding abstraction.

For OI button mapping to be done properly, I needed to change the
Xbox Controller library to the Joystick library, and use JoystickButton
objects for the gamepad buttons in IO.